Use POSIX IO functions, and add some broken sort-code
fread() have been producing an error for me, but printing "sucess" with
perror(), and I've not been motivated to dig into it.

The sort code for some reason skips over some of values at state.tokens,
i.e. it fails on the assertion with sizes. I have fucking no ideas.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>