wip: use range-for loop to move through windows
It will be reverted by the next commit, and retained for historical-only
purposes. The problem with this is that operator++ and …!= needs to be
added to Slice instead. But it's still lot of work, and I'm not sure it
worth complicating the struct — is it perhaps better to just make a
function and pass a lambda?

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>